+++ firebase-firestore/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - [feature] Added support for `first`, `last`, `arrayAgg`, and `arrayAggDistinct` Pipeline expressions.
   [#7893](https://github.com/firebase/firebase-android-sdk/pull/7893)
+- [feature] Added support for `rand` and `trunc` Pipeline expressions.
+  [#7886](https://github.com/firebase/firebase-android-sdk/pull/7886)
 - [feature] Added support for Pipeline expressions `arrayFirst`, `arrayFirstN`, `arrayLast`,
   `arrayLastN`, `arrayMinimum`, `arrayMinimumN`, `arrayMaximum`, `arrayMaximumN`, `arrayIndexOf`,
   `arrayLastIndexOf` and `arrayIndexOfAll`.